+++ src/routes/routes.ts
@@ -115,8 +115,8 @@ export class Routes extends RouteCollection {
                         router.post('/communication/ping', FirewallController, 'pingCommunication').name('fwclouds.firewalls.communication.ping');
                         router.post('/communication/info', FirewallController, 'infoCommunication').name('fwclouds.firewalls.communication.info');
                         router.post('/plugin',FirewallController,'installPlugin').name('fwclouds.firewalls.communication.installPlugin');
-                        router.post('/systemctl', FirewallController,'systemctlCommunication').name('fwclouds.firewalls.communication.systemctlCommunication')
                         router.prefix('/:firewall(\\d+)', (router:RouterParser) => {
+                            router.post('/systemctl', FirewallController,'systemctlCommunication').name('fwclouds.firewalls.communication.systemctlCommunication')
 
                             router.prefix('/policyRules', (router: RouterParser) => {
                                 router.get('/read', PolicyRuleController, 'read').name('fwclouds.firewalls.policyRules.read');