+++ CHANGELOG.md
@@ -4,6 +4,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased] - 2023-10-24
+### Added
+- Systemctl management API call.
+
+## [Unreleased] - 2023-10-24
+### Added
+- HAProxy plugin.
+
+## [Unreleased] - 2023-10-20
+### Added
+- Added query data to get Agent host information.
+
+
 ## [1.9.2] - 2023-06-07
 ### Added
 - `GET updates/type/pkg` API call. It will be used by FWCloud-UI to get the updates information in a FWCloud DEB/RPM packages based installation.